==> scripts/0_Exploration_vector_gpkg.ipynb <==
### **1. Paramètres**

# Importation des bibliothèques
import os 
import re
import numpy as np
import geopandas as gpd
import fiona
import matplotlib.pyplot as plt
import rasterio
from rasterio.features import rasterize
from rasterio.mask import mask


# Chargement des donnéés
path_vectors = "../data_brut/vector"
shps = fiona.listlayers(path_vectors)

print("Couches disponibles dans le dossier (shps) : ")
for shp in shps:
    print(shp)

# Charger les couches
dep47_brut = gpd.read_file(path_vectors, layer ='parcelles_dep47')
dep73_brut = gpd.read_file(path_vectors, layer ='parcelles_chautagne_dep73')
dep82_bb_brut = gpd.read_file(path_vectors, layer ='parcelles_borde_basse_dep82')
dep82_sp_brut = gpd.read_file(path_vectors, layer ='parcelles_gfa_st_pierre_dep82')
dep_10_brut = gpd.read_file(path_vectors, layer = "Peupliers_Dep10_32631")

### **2. Nettoyer le dep47**

# Faire une copie de la couche
dep47 = dep47_brut.copy()
print(dep47.columns)

#### **2.1 Sélectionner les colonnes**

# Sélectionnér les colonnes à conserver
col_dep47 = ['id',
             'Ann�e_pl',
             'Cultivars',
             'Densit�',
             'ID_peuplie',
             'comment',
             'geometry'
             ]

# Filtrer les colonnes pour garder seulement celles qui sont nécessaires
dep47 = dep47[col_dep47]

# Afficher les colonnes filtrées
print(dep47.columns)

#### **2.2 Renommer les colonnes**

# Créer un dictionnaire avec les nouveaux noms de colonnes
nom_col_dep47 = {'id' : 'id_47',
                 'Ann�e_pl' : 'annee_plan',
                 'Cultivars' : 'cultivar',
                 'Densit�' : 'densite',
                 'ID_peuplie' : 'id_peulplie'
                 }

# Renommer les colonnes
dep47.rename(columns=nom_col_dep47, inplace=True)
# Afficher les nouvelles colonnes pour vérification
print("Colonnes après nettoyage et renommage:")
print(dep47.columns)

#### **2.3 Sélection de parcelles avec des cultivars uniques**

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep47 = dep47['cultivar'].unique()
for unique in uniques_dep47:
    print(unique)

# Fonction pour vérifier si la valeur contient exactement un cultivar
def seul_cultivar(valeur):
    # Vérifie d'abord si la valeur est une chaîne de caractères
    if not isinstance(valeur, str):
        return False
    # Si la cellule est vide, contient 'nan', ou contient seulement des espaces, retourner False
    if not valeur or valeur.strip().lower() == "nan" or valeur.strip() == "":
        return False
    # Si la valeur contient une virgule, les mots 'melange', 'essai', 'divers', 'et', ou 'Vesten Dellinois', retourner False
    if ',' in valeur or re.search(r'\b(melange|essai|essais|divers|et)\b', valeur, re.IGNORECASE) or "vesten dellinois".lower() in valeur.lower():
        return False
    # Exclure 'Noyer/robinier', mas NÃO exclua 'I45/51'
    if re.search(r'noyer/robinier', valeur.lower()):
        return False
    # Si la valeur contient des numéros au format d'essai (ex. 1000-1, 0685-16), retourner False
    if re.search(r'\d{3,4}-\d{1,2}', valeur):
        return False
    # Supprime les espaces superflus et convertit en minuscules
    valeur = valeur.strip().lower()
    # Divise la chaîne sur des tirets (avec ou sans espace) ou des espaces
    cultivars = re.split(r'\s*-\s*', valeur)
    # Vérifie s'il y a exactement un cultivar unique
    return len(set(cultivars)) == 1


# Filtrer les lignes avec un seul cultivar dans la colonne 'cultivar'
dep47_cultivar_unique = dep47[dep47['cultivar'].apply(seul_cultivar)]

# Affiche les premières lignes pour vérifier le résultat
print("Nombre total de lignes avant filtrage :", len(dep47))
print("Nombre de lignes avec un seul cultivar :", len(dep47_cultivar_unique))

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep47_single_cultivar = dep47_cultivar_unique['cultivar'].unique()
for unique_clean in uniques_dep47_single_cultivar:
    print(unique_clean)
    
# Afficher les premières lines
dep47_cultivar_unique.head()

### **3. Nettoyer le dep73**

# Faire une copie de la couche
dep73 = dep73_brut.copy()
print(dep73.columns)
dep73.head(2)

#### **3.1 Sélectionner les colonnes**

# Sélectionnér les colonnes à conserver
col_dep73 = ['fid',
             'Annee_plan',
             'Cultivar1',
             'geometry'
             ]

# Filtrer les colonnes pour garder seulement celles qui sont nécessaires
dep73 = dep73[col_dep73]

# Afficher les colonnes filtrées
print(dep73.columns)

#### **3.2 Renommer les colonnes**

# Créer un dictionnaire avec les nouveaux noms de colonnes
nom_col_dep73 = {'fid' : 'id_73',
                 'Annee_plan' : 'annee_plan',
                 'Cultivar1' : 'cultivar'
                 }

# Renommer les colonnes
dep73.rename(columns=nom_col_dep73, inplace=True)

# Afficher les nouvelles colonnes pour vérification
print("Colonnes après nettoyage et renommage:")
print(dep73.columns)

#### **3.3 Sélection de parcelles avec des cultivars uniques**

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep73 = dep73['cultivar'].unique()
for unique in uniques_dep73:
    print(unique)

# Filtrer les lignes avec un seul cultivar dans la colonne 'cultivars'
dep73_cultivar_unique = dep73[dep73['cultivar'].apply(seul_cultivar)]

# Affiche les premières lignes pour vérifier le résultat
print("Nombre total de lignes avant filtrage :", len(dep73))
print("Nombre de lignes avec un seul cultivar :", len(dep73_cultivar_unique))

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep73_single_cultivar = dep73_cultivar_unique['cultivar'].unique()
for unique_73clean in uniques_dep73_single_cultivar:
    print(unique_73clean)
    
# Afficher les premières lines
dep73_cultivar_unique.head()

### **4. Nettoyer le dep82 - Borde Basse**

# Faire une copie de la couche de la région Borde Basse
dep82_bb = dep82_bb_brut.copy()
print(dep82_bb.columns)
dep82_bb.head(2)

#### **41. Sélectionner les colonnes dep82 - Borde Basse**

# Sélectionnér les colonnes à conserver
col_dep82_bb = ['ID_user',
                'Essence',
                'Année_pla',
                'geometry'
                ]

# Filtrer les colonnes pour garder seulement celles qui sont nécessaires
dep82_bb = dep82_bb[col_dep82_bb]

# Afficher les colonnes filtrées
print(dep82_bb.columns)
dep82_bb.head(2)

#### **4.2 Renommer les colonnes**

# Créer un dictionnaire avec les nouveaux noms de colonnes
nom_col_dep82_bb = {'ID_user' : 'id_82_bb',
                 'Année_pla' : 'annee_plan',
                 'Essence' : 'cultivar'
                 }

# Renommer les colonnes
dep82_bb.rename(columns=nom_col_dep82_bb, inplace=True)

# Afficher les nouvelles colonnes pour vérification
print("Colonnes après nettoyage et renommage:")
print(dep82_bb.columns)

#### **4.3 Sélection des parcelles avec une seul cultivar**

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep82_bb = dep82_bb['cultivar'].unique()
for unique_82_bb in uniques_dep82_bb:
    print(unique_82_bb)

# Filtrer les lignes avec un seul cultivar dans la colonne 'cultivar'
dep82_bb_cultivar_unique = dep82_bb[dep82_bb['cultivar'].apply(seul_cultivar)]

# Affiche les premières lignes pour vérifier le résultat
print("Nombre total de lignes avant filtrage :", len(dep82_bb))
print("Nombre de lignes avec un seul cultivar :", len(dep82_bb_cultivar_unique))

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep82_bb_single_cultivar = dep47_cultivar_unique['cultivar'].unique()
for unique_dep82_bb_single_cultivar in uniques_dep82_bb_single_cultivar:
    print(unique_dep82_bb_single_cultivar)
    
# Afficher les premières lines
dep82_bb_cultivar_unique.head()

### **5. Nettoyer le dep82 - Saint Pierre**

# Faire une copie de la couche de la région Borde Basse
dep82_sp = dep82_sp_brut.copy()
print(dep82_sp.columns)
dep82_sp.head(2)

#### **5.1 Sélectionner les colonnes**

# Sélectionnér les colonnes à conserver
col_dep82_sp = ['ID_user',
                'Espece',
                'Annee',
                'geometry'
                ]

# Filtrer les colonnes pour garder seulement celles qui sont nécessaires
dep82_sp = dep82_sp[col_dep82_sp]

# Afficher les colonnes filtrées
print(dep82_sp.columns)
dep82_sp.head(3)

#### **5.2 Renommer les colonnes**

# Créer un dictionnaire avec les nouveaux noms de colonnes
nom_col_dep82_sp = {'ID_user' : 'id_82_sp',
                  'Annee' : 'annee_plan',
                 'Espece' : 'cultivar'
                 }

# Renommer les colonnes
dep82_sp.rename(columns=nom_col_dep82_sp, inplace=True)

# Afficher les nouvelles colonnes pour vérification
print("Colonnes après nettoyage et renommage:")
print(dep82_sp.columns)

#### **5.3 Sélection des parcelles**

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep82_sp = dep82_sp['cultivar'].unique()
for unique_82_sp in uniques_dep82_sp:
    print(unique_82_sp)

# Filtrer les lignes avec un seul cultivar dans la colonne 'cultivar'
dep82_sp_cultivar_unique = dep82_sp[dep82_sp['cultivar'].apply(seul_cultivar)]

# Affiche les premières lignes pour vérifier le résultat
print("Nombre total de lignes avant filtrage :", len(dep82_sp))
print("Nombre de lignes avec un seul cultivar :", len(dep82_sp_cultivar_unique))

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep82_sp_single_cultivar = dep82_sp_cultivar_unique['cultivar'].unique()
for unique_dep82_sp_single_cultivar in uniques_dep82_sp_single_cultivar:
    print(unique_dep82_sp_single_cultivar)
    
# Afficher les premières lines
dep82_sp_cultivar_unique.head()

### **6. Nettoyer le dep10**

# Faire une copie de la couche de département 10
dep10 = dep_10_brut.copy()
print(dep10.columns)
dep10.head(2)

#### **6.1. Sélectionner les colonnes de dep10**

# Sélectionnér les colonnes à conserver
col_dep10 = ['ID',
                'd_essenc_1',
                'Annee',
                'geometry'
                ]

# Filtrer les colonnes pour garder seulement celles qui sont nécessaires
dep10 = dep10[col_dep10]

# Afficher les colonnes filtrées
print(dep10.columns)
dep10.head(3)

#### **6.2 Renommer les colonnes**

# Créer un dictionnaire avec les nouveaux noms de colonnes
nom_col_dep10 = {'ID' : 'id_10',
                  'Annee' : 'annee_plan',
                 'd_essenc_1' : 'cultivar'
                 }

# Renommer les colonnes
dep10.rename(columns=nom_col_dep10, inplace=True)

# Afficher les nouvelles colonnes pour vérification
print("Colonnes après nettoyage et renommage:")
print(dep10.columns)

#### **6.3 Sélection des parcelles**

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep10 = dep10['cultivar'].unique()
for unique_dep10 in uniques_dep10:
    print(unique_dep10)

# Filtrer les lignes avec un seul cultivar dans la colonne 'cultivar'
dep10_cultivar_unique = dep10[dep10['cultivar'].apply(seul_cultivar)]

# Affiche les premières lignes pour vérifier le résultat
print("Nombre total de lignes avant filtrage :", len(dep10))
print("Nombre de lignes avec un seul cultivar :", len(dep10_cultivar_unique))

# Boucle pour afficher les valeurs uniques de la colonne 'cultivars'
uniques_dep10_single_cultivar = dep10_cultivar_unique['cultivar'].unique()
for unique_dep10_single_cultivar in uniques_dep10_single_cultivar:
    print(unique_dep10_single_cultivar)
    
# Afficher les premières lines
dep10_cultivar_unique.head()
