==> scripts/2_Decoupe_rasters.ipynb <==
### **1. Paramètres**

# Importation des bibliothèques
import os

# Importation des fonctions
from functions_decoupe import (add_band_names, create_vrt, clip_and_align_raster, clip_raster, verifier_raster)

# Définition des années et des zones d'intérêt
annees = [2017, 2018, 2019, 2020, 2021, 2022]
zones = ['T30TYP', 'T30TYQ', 'T31TCJ', 'T31TGL', 'T31UEP']

# Liste des métriques LiDAR à traiter
lidar_metrics = ['grid_CC', 'grid_ENL', 'grid_MOCH', 'grid_PAI', 'grid_VCI']

# Définition des chemins vers les répertoires de données
base_dir = "../data_brut/raster"
output_confidence_temp = "../data_brut/raster/stacks/confidence"  # Répertoire temporaire pour les VRT de confiance
output_lidar_temp = "../data_brut/raster/stacks/lidar"           # Répertoire temporaire pour les rasters LiDAR découpés
output_confidence_final = "../data_final/raster/confidence"      # Répertoire final pour les rasters de confiance découpés
output_lidar_final = "../data_final/raster/lidar"                # Répertoire final pour les rasters LiDAR découpés

# Chemin vers le GeoPackage utilisé pour la découpage
gpkg_path = '../data_final/vector/peupleraies_lidar_parcelle.gpkg'

### **2. Créer les piles des rasters de confidence**

# 2.1 Création des VRT pour les rasters de confiance
for zone in zones:
    create_vrt(base_dir, zone, annees, 'confidence', output_confidence_temp)

# 2.2 Découpage des VRT de confiance et sauvegarde au format TIFF
for zone in zones:
    vrt_path = os.path.join(output_confidence_temp, f'confidence_stack_{zone}.vrt')
    output_raster = clip_raster(
        raster_path=vrt_path,
        gpkg_path=gpkg_path,
        output_dir=output_confidence_final,
        output_name=f"confidence_clipped_{zone}",
        nodata_value=-999,
        dtype_value='int16'
    )
    
    # Ajout des noms des bandes correspondant aux années
    add_band_names(output_raster, [str(annee) for annee in annees])

### **3. Créer la pile des métriques lidar**

# 3.1 Initialisation de la liste pour stocker les rasters LiDAR découpés
lidar_rasters_clipped = []

# Création des répertoires de sortie si ils n'existent pas
os.makedirs(output_lidar_temp, exist_ok=True)
os.makedirs(output_lidar_final, exist_ok=True)

# Itération sur chaque zone pour traiter les métriques LiDAR
for zone in zones:
    if zone == "T31UEP":  # Ignorer la tuile sans métriques LiDAR
        continue
    
    # Chemin vers le raster de référence (confiance découée) pour la zone actuelle
    reference_clip_path = os.path.join(output_confidence_final, f"confidence_clipped_{zone}.tif")
    
    for metric in lidar_metrics:
        input_raster = os.path.join(base_dir, f"{metric}.tif")
        output_clipped = os.path.join(output_lidar_temp, f"{metric}_clipped_{zone}.tif")
        
        if os.path.exists(input_raster):
            # 3.1.1 Aligner et clipper le raster LiDAR
            clip_and_align_raster(input_raster, reference_clip_path, output_clipped)
            
            # 3.1.2 Découper le raster LiDAR aligné et clipé
            output_cliped = clip_raster(
                raster_path=output_clipped,
                gpkg_path=gpkg_path,
                output_dir=output_lidar_final,
                output_name=f"{metric}_clipped_{zone}",
                nodata_value=-999,
                dtype_value='float32'
            )
            
            # Ajouter le raster LiDAR découpé à la liste
            lidar_rasters_clipped.append(output_cliped)
        else:
            print(f" Fichier LiDAR manquant : {input_raster} pour la tuile {zone}")


### **4. Vérification des raster générés**

# 4.1 Définition des chemins des rasters de confiance à vérifier
stack_confidence = [
    os.path.join(output_confidence_final, 'confidence_clipped_T30TYP.tif'),
    os.path.join(output_confidence_final, 'confidence_clipped_T30TYQ.tif'),
    os.path.join(output_confidence_final, 'confidence_clipped_T31TCJ.tif'),
    os.path.join(output_confidence_final, 'confidence_clipped_T31TGL.tif'),
    os.path.join(output_confidence_final, 'confidence_clipped_T31UEP.tif')
]

# 4.2 Définition du chemin du raster LiDAR à vérifier
lidar = lidar_rasters_clipped  # Liste des rasters LiDAR découpé

# 4.3 Vérification des statistiques et des valeurs NoData pour chaque raster de confiance
for raster in stack_confidence:
    verifier_raster(raster)   

# 4.4 Vérification des statistiques et des valeurs NoData pour chaque raster LiDAR
for raster in lidar:
    verifier_raster(raster)
