==> scripts/5_Boxnotch_confidenceXage.ipynb <==
### **1. Paramètres**

# Importation des bibliothèques
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from matplotlib.backends.backend_pdf import PdfPages

from functions_plots import top_cultivars, boxnotch_confidenceXage, grid_boxnotch_confidenceXage_par_anne

# Chemin vers les fichiers
csv_path = '../data_final/tableaux/'
output_path = '../results/1_confidenceXage'
output_path_par_annee = '../results/2_confidenceXage_par_annee'

# Vérifier si le dossier de sortie existe ; le créer si nécessaire
if not os.path.exists(output_path):
    os.makedirs(output_path)

if not os.path.exists(output_path_par_annee):
    os.makedirs(output_path_par_annee)

# Lire le fichier CSV
df_px_filtre_dept10 = pd.read_csv(os.path.join(csv_path, 'df_pixel_filtre_dept10.csv'), dtype={'id_parc': str})

### **2. Faire les boxnotchs pour tous cultivars confundus**

# Obtenir toutes les valeurs uniques de la colonne 'source' dans le DataFrame filtré
unique_sources = df_px_filtre_dept10['source'].unique()

# Générer une palette de couleurs avec 'tab10' pour chaque valeur unique de 'source'
source_colors = plt.cm.tab10(np.linspace(0, 1, len(unique_sources)))
source_color_map = dict(zip(unique_sources, source_colors))

# Filtre pour les 12 premières années
df_px_filtre_dept10_all = df_px_filtre_dept10[(df_px_filtre_dept10['age_plan'] >= 1) & (df_px_filtre_dept10['age_plan'] <= 12)]

# Création d'un fichier PDF combinant tous les boxplots déjà existants
pdf_path_all = os.path.join(output_path, "Boxnotch_confidenceXage_all.pdf")
with PdfPages(pdf_path_all) as pdf:
    boxnotch_confidenceXage(
        df=df_px_filtre_dept10,  # Utiliser tout le DataFrame
        cultivar=None,  # Spécifier que c'est pour tous les cultivars combinés
        index=0,  # Indiquer un index fixe pour tous les cultivars
        output_path=output_path,
        color_map=source_color_map,
        pdf=pdf
    )

print(f"PDF pour tous les cultivars combinés sauvegardé : {pdf_path_all}")

### **3. Faire les boxnotchs par cultivar**

# Faire un dicitonnaires avec les cultivars principaux
dict_top_cultivars = top_cultivars(df_px_filtre_dept10)
print(f"Nombre de cultivars sélectionnés: {len(dict_top_cultivars)}")

# Imprimer les cultivars triés par nombre de pixels
for cultivar, count in dict_top_cultivars.items():
    print(f"{cultivar}: {count} pixels")

# Filtrer le DataFrame pour les cultivars principaux
df_px_filtre_dept10_top = df_px_filtre_dept10[
    df_px_filtre_dept10['cultivar_n'].isin(dict_top_cultivars)
]

# Générer les boxplots pour chaque cultivar
pdf_path_top = os.path.join(output_path, "Boxnotch_confidenceXage_top_cultivars.pdf")
with PdfPages(pdf_path_top) as pdf:
    for index, cultivar in enumerate(dict_top_cultivars, start=1):
        boxnotch_confidenceXage(
            df=df_px_filtre_dept10_top,
            cultivar=cultivar,
            index=index,  # Utiliser l'index basé sur le cultivar
            output_path=output_path,
            color_map=source_color_map,
            pdf=pdf
        )
print(f"PDF pour les cultivars principaux sauvegardé : {pdf_path_top}")

### **4. Faire le grid pour tracer le boxnotchs par cultivar par année**

# Définir une palette de couleurs pour chaque année
color_palette = {
    2017: "skyblue", 2018: "orange", 2019: "green",
    2020: "red", 2021: "purple", 2022: "brown"
}

# Création d'un fichier PDF combinant tous les boxplots
with PdfPages(os.path.join(output_path_par_annee, "Boxnotch_confidenceXage_par_annee.pdf")) as pdf:
    for index, cultivar in enumerate(dict_top_cultivars, start=1):
        grid_boxnotch_confidenceXage_par_anne(df=df_px_filtre_dept10_top,
                                              cultivar=cultivar,
                                              output_path=output_path_par_annee,
                                              color_palette=color_palette,
                                              index=index,
                                              pdf=pdf)

print("PDF combiné avec tous les boxplots sauvegardé avec succès.")
