==> scripts/6_Boxnotch_lidar.ipynb <==
### **1. Paramètres**

# Importation des bibliothèques
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from matplotlib.backends.backend_pdf import PdfPages

from functions_plots import top_cultivars, boxnotch_lidar_metrics, boxnotch_confidenceXage_lidar_metrics

# Chemin vers les fichiers
csv_path = '../data_final/tableaux/'
output_path = '../results/3_lidar_metrics'
output_path_grid = '../results/4_confidenceXage_lidar_metrics'

# Vérifier si le dossier de sortie existe ; le créer si nécessaire
if not os.path.exists(output_path):
    os.makedirs(output_path)

if not os.path.exists(output_path_grid):
    os.makedirs(output_path_grid)

# Lire le fichier CSV
df_px_filtre_lidar = pd.read_csv(os.path.join(csv_path, 'df_pixel_filtre_lidar.csv'), dtype={'id_parc': str})

### **2. Faire les boxplots pour tous cultivars confundus**

# Obtenir toutes les valeurs uniques de la colonne 'source' dans le DataFrame filtré
unique_sources = df_px_filtre_lidar['source'].unique()

# Générer une palette de couleurs avec 'tab10' pour chaque valeur unique de 'source'
source_colors = plt.cm.tab10(np.linspace(0, 1, len(unique_sources)))
source_color_map = dict(zip(unique_sources, source_colors))

metrics = ['grid_CC', 'grid_PAI', 'grid_ENL', 'grid_VCI', 'grid_MOCH']

# Calculer les limites Y pour chaque métrique
y_limits = {}
for metric in metrics:
    y_min = df_px_filtre_lidar[metric].min()
    y_max = df_px_filtre_lidar[metric].max()
    y_limits[metric] = (y_min, y_max)

# Filtre pour les 12 premières années
df_px_filtre_lidar_all = df_px_filtre_lidar[(df_px_filtre_lidar['age_plan'] >= 1) & (df_px_filtre_lidar['age_plan'] <= 12)]

# Générer les graphiques
pdf_path_all = os.path.join(output_path_grid, "Boxnotch_confidenceXage_lidar_metrics_all.pdf")
with PdfPages(pdf_path_all) as pdf:
    for metric in metrics:
            boxnotch_confidenceXage_lidar_metrics(
                df=df_px_filtre_lidar_all,
                metric=metric,
                color_map=source_color_map,
                y_limits=y_limits[metric],
                pdf=pdf,
                output_path=output_path_grid,
                cultivar_index=None  # Indique que tout le CSV est utilisé
            )

print(f"PDF avec le graphique des tous cultivars confundus sauvegardé à: {pdf_path_all}")

### **3. Faire les boxnotchs pour les métriques**

# Faire un dicitonnaires avec les cultivars principaux
dict_top_cultivars = top_cultivars(df_px_filtre_lidar, max_age=12, min_years=2, top_n=20)
print(f"Nombre de cultivars sélectionnés: {len(dict_top_cultivars)}")

# Imprimer les cultivars triés par nombre de pixels
for cultivar, count in dict_top_cultivars.items():
    print(f"{cultivar}: {count} pixels")

# Prendre juste les tops cultivar  pour faire les boxplots
df_px_filtre_lidar_top = df_px_filtre_lidar[df_px_filtre_lidar['cultivar_n'].isin(dict_top_cultivars)]
df_px_filtre_lidar_top = df_px_filtre_lidar_top[(df_px_filtre_lidar_top['age_plan'] >= 1) & (df_px_filtre_lidar_top['age_plan'] <= 12)]
df_px_filtre_lidar_top['age_plan'] = df_px_filtre_lidar_top['age_plan'].astype(int)

# Générer les graphiques
pdf_path = os.path.join(output_path, "Boxnotch_lidar_metrics_top_cultivars.pdf")
with PdfPages(pdf_path) as pdf:
    for cultivar_index, (cultivar, _) in enumerate(dict_top_cultivars.items(), start=1):
        for metric in metrics:
            boxnotch_lidar_metrics(
                df=df_px_filtre_lidar_top,
                cultivar=cultivar,
                metric=metric,
                color_map=source_color_map,
                y_limits=y_limits[metric],
                pdf=pdf,
                output_path=output_path,
                cultivar_index=cultivar_index
            )

print(f"PDF avec tous les graphiques sauvegardé à: {pdf_path}")


### **4. Faire le grid des boxnotchs : confidece + métriques**

pdf_path_grid = os.path.join(output_path_grid, "Boxnotch_confidenceXage_lidar_metrics_top_cultivars.pdf")
with PdfPages(pdf_path_grid) as pdf:
    for cultivar_index, (cultivar, _) in enumerate(dict_top_cultivars.items(), start=1):
        for metric in metrics:
            boxnotch_confidenceXage_lidar_metrics(
                df=df_px_filtre_lidar_top,
                cultivar=cultivar,
                metric=metric,
                color_map=source_color_map,
                y_limits=y_limits[metric],
                pdf=pdf,
                output_path=output_path_grid,
                cultivar_index=cultivar_index  # Index pour cultivar
            )

print(f"PDF avec tous les graphiques sauvegardé à: {pdf_path_grid}")
